+++ packages/moonstone/IncrementSlider/tests/IncrementSlider-specs.js
@@ -203,10 +203,21 @@ describe('IncrementSlider Specs', () => {
 
 	it('should set decrementButton "aria-label" to value and hint string', function () {
 		const incrementSlider = mount(
-			<IncrementSlider value={0} />
+			<IncrementSlider value={10} />
 		);
 
-		const expected = '0 press ok button to decrease the value';
+		const expected = '10 press ok button to decrease the value';
+		const actual = incrementSlider.find(`.${css.decrementButton}`).prop('aria-label');
+
+		expect(actual).to.equal(expected);
+	});
+
+	it('should not set decrementButton "aria-label" when decrementButton is disabled', function () {
+		const incrementSlider = mount(
+			<IncrementSlider disabled value={10} />
+		);
+
+		const expected = null;
 		const actual = incrementSlider.find(`.${css.decrementButton}`).prop('aria-label');
 
 		expect(actual).to.equal(expected);
@@ -214,10 +225,21 @@ describe('IncrementSlider Specs', () => {
 
 	it('should set incrementButton "aria-label" to value and hint string', function () {
 		const incrementSlider = mount(
-			<IncrementSlider value={0} />
+			<IncrementSlider value={10} />
 		);
 
-		const expected = '0 press ok button to increase the value';
+		const expected = '10 press ok button to increase the value';
+		const actual = incrementSlider.find(`.${css.incrementButton}`).prop('aria-label');
+
+		expect(actual).to.equal(expected);
+	});
+
+	it('should not set incrementButton "aria-label" when incrementButton is disabled', function () {
+		const incrementSlider = mount(
+			<IncrementSlider disabled value={10} />
+		);
+
+		const expected = null;
 		const actual = incrementSlider.find(`.${css.incrementButton}`).prop('aria-label');
 
 		expect(actual).to.equal(expected);