+++ packages/moonstone/Scroller/ScrollButton.js
@@ -1,8 +1,9 @@
-import Holdable from '@enact/ui/Holdable';
+import Holdable from '../internal/Holdable';
 import kind from '@enact/core/kind';
 import onlyUpdateForKeys from 'recompose/onlyUpdateForKeys';
-import React from 'react';
 import PropTypes from 'prop-types';
+import React from 'react';
+import Toggleable from '@enact/ui/Toggleable';
 
 import $L from '../internal/$L';
 import IconButton from '../IconButton';
@@ -49,6 +50,15 @@ const ScrollButtonBase = kind({
 		direction: PropTypes.oneOf(['down', 'left', 'right', 'up']).isRequired,
 
 		/**
+		 * When `true`, the `aria-label` is set.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		active: PropTypes.bool,
+
+		/**
 		 * When `true`, the component is shown as disabled and does not generate `onClick`
 		 * [events]{@glossary event}.
 		 *
@@ -64,11 +74,12 @@ const ScrollButtonBase = kind({
 	},
 
 	computed: {
-		'aria-label': ({direction}) => $L(`scroll ${direction}`),
+		'aria-label': ({active, direction}) => active ? $L(`scroll ${direction}`) : null,
 		className: ({direction, styler}) => styler.append(classNameMap[direction])
 	},
 
 	render: ({children, disabled, ...rest}) => {
+		delete rest.active;
 		delete rest.direction;
 
 		return (
@@ -96,7 +107,10 @@ const ScrollButtonBase = kind({
  */
 const ScrollButton = withSkinnableProps(
 	onlyUpdateForKeys(['children', 'disabled', 'skin'])(
-		ScrollButtonBase
+		Toggleable(
+			{activate: 'onFocus', deactivate: 'onBlur', toggle: null},
+			ScrollButtonBase
+		)
 	)
 );