+++ packages/spotlight/CHANGELOG.md
@@ -12,6 +12,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Changed
 
+- `spotlight/Spottable` to remove focus from a component when it becomes disabled and move it to another component if not explicitly moved during the `onSpotlightDisappear` event callback
+
 ### Fixed
 
 - Navigating through spottable components while holding down a directional key