+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [1.4.0] - 2017-06-29
+
+### Added
+
+- `spotlight` handlers for window focus events
+
+### Fixed
+
+- `spotlight` navigation through spottable components while holding down a directional key
+- `spotlight` support for preventing 5-way navigation out of a container using an empty selector
+- `spotlight` container support for default elements within subcontainers
+
 ## [1.3.1] - 2017-06-14
 
 ### Fixed